+++ test/wadogo/scale/linear_test.clj
@@ -40,9 +40,20 @@
       (is (:pass? (tc/quick-check
                    100
                    (prop/for-all* [(gen/double* {:min 0.0 :max 100.0 :infinite? false :NaN? false})]
-                                  (fn [n] (and (m/approx-eq (l n) (* 2.0 n) 6)
-                                              (m/approx-eq (s/forward l n) (* 2.0 n) 6)
-                                              (m/approx-eq (s/inverse l n) (* 0.5 n) 6))))))))))
+                                  (fn [n] (let [acc 6
+                                               n (m/approx n acc)
+                                               f1 (m/approx (l n) acc)
+                                               f2 (m/approx (s/forward l n) acc)
+                                               i (m/approx (s/inverse l n) acc)
+                                               fres (m/approx (* 2.0 n) acc)
+                                               ires (m/approx (* 0.5 n) acc)
+                                               res (and (== f1 f2 fres)
+                                                        (== i ires))]
+                                           #_(when-not res (println [(s/forward l n)
+                                                                     (s/inverse l n)
+                                                                     (* 2.0 n) (* 0.5 n)]
+                                                                    [f1 f2 i fres ires]))
+                                           res)))))))))
 
 (deftest zero-length-domain-or-range
   (let [l (s/scale :linear {:domain [0 0]})]